feat(linter): cleanup json converter logic
